+++ 3dTerrainGeneration/rendering/Mesh.cs
@@ -9,9 +9,9 @@ namespace _3dTerrainGeneration.rendering
     internal struct Mesh
     {
         public int Width, Height;
-        public ushort[][] Data;
+        public uint[][] Data;
 
-        public Mesh(int width, int height, ushort[][] data) : this()
+        public Mesh(int width, int height, uint[][] data) : this()
         {
             Width = width;
             Height = height;
