+++ 3dTerrainGeneration/rendering/MeshData.cs
@@ -19,9 +19,9 @@ namespace _3dTerrainGeneration.rendering
 
         public int Width, Height;
 
-        public ushort[] MeshSingle(byte emission, int scale = 1)
+        public uint[] MeshSingle(byte emission, int scale = 1)
         {
-            ushort[][] mesh = Mesh(emission, scale);
+            uint[][] mesh = Mesh(emission, scale);
 
             int len = 0;
             for (int i = 0; i < 6; i++)
@@ -29,7 +29,7 @@ namespace _3dTerrainGeneration.rendering
                 len += mesh[i].Length;
             }
 
-            ushort[] meshSingle = new ushort[len];
+            uint[] meshSingle = new uint[len];
             int index = 0;
             for (int i = 0; i < 6; i++)
             {
@@ -40,7 +40,7 @@ namespace _3dTerrainGeneration.rendering
             return meshSingle;
         }
 
-        public ushort[][] Mesh(byte emission, int scale = 1)
+        public uint[][] Mesh(byte emission, int scale = 1)
         {
             if(blocks == null)
             {
